refactor: move provider with new namespacing
