+++ vda5050_execution/include/vda5050_execution/core/provider.hpp
@@ -16,17 +16,20 @@
  * limitations under the License.
  */
 
-#ifndef VDA5050_EXECUTION__PROVIDER_HPP_
-#define VDA5050_EXECUTION__PROVIDER_HPP_
+#ifndef VDA5050_EXECUTION__CORE__PROVIDER_HPP_
+#define VDA5050_EXECUTION__CORE__PROVIDER_HPP_
 
 #include <functional>
 #include <memory>
+#include <mutex>
 #include <unordered_map>
 #include <utility>
 
-#include "vda5050_execution/update.hpp"
+#include "vda5050_execution/core/base.hpp"
 
 namespace vda5050_execution {
+
+namespace core {
 
 class Provider : public std::enable_shared_from_this<Provider>
 {
@@ -38,10 +41,11 @@ public:
       std::is_base_of_v<UpdateBase, UpdateT>,
       "Update must be derived from UpdateBase");
 
-    push_shared(std::make_shared<UpdateT>(std::forward<Args>(args)...));
+    auto update = std::make_shared<UpdateT>(std::forward<Args>(args)...);
+    push_shared(std::move(update));
   }
 
-  void push_shared(std::shared_ptr<UpdateBase> update) const;
+  void push_shared(std::shared_ptr<UpdateBase> update);
 
   template <typename UpdateT>
   void on(std::function<void(std::shared_ptr<UpdateT>)> provider)
@@ -55,14 +59,17 @@ public:
       cb(std::static_pointer_cast<UpdateT>(update));
     };
 
-    providers_[std::type_index(typeid(UpdateT))] = std::move(wrapper);
+    std::lock_guard<std::mutex> lock(registry_mutex_);
+    providers_[std::type_index(typeid(UpdateT))].push_back(std::move(wrapper));
   }
 
 private:
   using ErasedProvider = std::function<void(std::shared_ptr<UpdateBase>)>;
-  std::unordered_map<std::type_index, ErasedProvider> providers_;
+  std::unordered_map<std::type_index, std::vector<ErasedProvider>> providers_;
+  std::mutex registry_mutex_;
 };
 
+}  // namespace core
 }  // namespace vda5050_execution
 
-#endif  // VDA5050_EXECUTION__PROVIDER_HPP_
+#endif  // VDA5050_EXECUTION__CORE__PROVIDER_HPP_
